Filed DummyItemMaker code in trait
